+++ src/main/java/learn/thyme/thyteams/team/PlayerPosition.java
@@ -1,0 +1,9 @@
+package learn.thyme.thyteams.team;
+
+public enum PlayerPosition {
+    POINT_GUARD,
+    SHOOTING_GUARD,
+    SMALL_FORWARD,
+    POWER_FORWARD,
+    CENTER
+}