+++ src/main/java/learn/thyme/thyteams/team/TeamPlayerId.java
@@ -1,0 +1,15 @@
+package learn.thyme.thyteams.team;
+
+import io.github.wimdeblauwe.jpearl.AbstractEntityId;
+
+import java.util.UUID;
+
+public class TeamPlayerId extends AbstractEntityId<UUID> {
+
+    protected TeamPlayerId() {
+    }
+
+    public TeamPlayerId(UUID id) {
+        super(id);
+    }
+}