+++ src/main/java/learn/thyme/thyteams/team/TeamRepositoryCustom.java
@@ -2,4 +2,6 @@ package learn.thyme.thyteams.team;
 
 public interface TeamRepositoryCustom {
     TeamId nextId();
+
+    TeamPlayerId nextPlayerId();
 }