+++ src/main/java/learn/thyme/thyteams/team/TeamService.java
@@ -10,15 +10,17 @@ import java.util.Optional;
 public interface TeamService {
     Page<TeamSummary> getTeams(Pageable pageable);
 
-    Team createTeam(String name, User coach);
-
-    Team createTeam(String name, UserId coachId);
+    Team createTeam(CreateTeamParameters parameters);
 
     Optional<Team> getTeam(TeamId teamId);
 
-    Team editTeam(TeamId teamId, long version, String name, UserId coachId);
+    Optional<Team> getTeamWithPlayers(TeamId teamId);
+
+    Team editTeam(TeamId teamId, EditTeamParameters parameters);
 
     void deleteTeam(TeamId teamId);
 
     void deleteAllTeams();
+
+    Team addPlayer(TeamId id, long version, UserId userId, PlayerPosition position);
 }