+++ src/main/java/learn/thyme/thyteams/team/web/CreateTeamFormData.java
@@ -1,10 +1,16 @@
 package learn.thyme.thyteams.team.web;
 
+import learn.thyme.thyteams.team.CreateTeamParameters;
+import learn.thyme.thyteams.team.TeamPlayerParameters;
 import learn.thyme.thyteams.user.UserId;
 
+import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import java.util.Arrays;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 public class CreateTeamFormData {
     @NotBlank
@@ -12,6 +18,15 @@ public class CreateTeamFormData {
     private String name;
     @NotNull
     private UserId coachId;
+
+    @NotNull
+    @Size(min = 1)
+    @Valid
+    private TeamPlayerFormData[] players;
+
+    public CreateTeamFormData() {
+        players = new TeamPlayerFormData[]{new TeamPlayerFormData()};
+    }
 
     public String getName() {
         return name;
@@ -27,5 +42,23 @@ public class CreateTeamFormData {
 
     public void setCoachId(UserId coachId) {
         this.coachId = coachId;
+    }
+
+    public TeamPlayerFormData[] getPlayers() {
+        return players;
+    }
+
+    public void setPlayers(TeamPlayerFormData[] players) {
+        this.players = players;
+    }
+
+    public CreateTeamParameters toParameters() {
+        return new CreateTeamParameters(name, coachId, getTeamPlayerParameters());
+    }
+
+    protected Set<TeamPlayerParameters> getTeamPlayerParameters() {
+        return Arrays.stream(players)
+                .map(tpFormData -> new TeamPlayerParameters(tpFormData.getPlayerId(), tpFormData.getPosition()))
+                .collect(Collectors.toSet());
     }
 }